Refactor Translation Menu
